[LINUXEP-7604] [LINUXEP-7497] concurrency fixes for repatchFilesystem.
Fix several possible race conditions in repatchFilesystem, by taking a write lock on mPatchLock and re-fetching the patch to be sure it's still valid.
Don't getPatch() / putPatch() for every item while iterating patches, as we already hold a read lock on mPatchLock.
Removed patch->lock as changes are now always made with a write lock on GL_object.mPatchLock held.

[+DLCL]

[git-p4: depot-paths = "//dev/talpa/main/": change = 2808995]